+++ src/app/components/Inputs/SmartInput/index.tsx
